+++ flash/core/data/data_pipeline.py
@@ -179,7 +179,8 @@ class DataPipeline:
         if object_type is None:
             object_type = Preprocess
 
-        prefixes = ['']
+        prefixes = []
+
         if stage in (RunningStage.TRAINING, RunningStage.TUNING):
             prefixes += ['train', 'fit']
         elif stage == RunningStage.VALIDATING:
@@ -189,9 +190,11 @@ class DataPipeline:
         elif stage == RunningStage.PREDICTING:
             prefixes += ['predict']
 
+        prefixes += [None]
+
         for prefix in prefixes:
             if cls._is_overriden(function_name, process_obj, object_type, prefix=prefix):
-                return f'{prefix}_{function_name}'
+                return function_name if prefix is None else f'{prefix}_{function_name}'
 
         return function_name