+++ flash/core/data/data_source.py
@@ -36,6 +36,7 @@ import torch
 from pytorch_lightning.trainer.states import RunningStage
 from pytorch_lightning.utilities.enums import LightningEnum
 from torch.nn import Module
+from torch.utils.data.dataset import Dataset
 
 from flash.core.data.auto_dataset import AutoDataset, BaseAutoDataset, IterableAutoDataset
 from flash.core.data.properties import ProcessState, Properties
@@ -143,6 +144,7 @@ class DefaultDataSources(LightningEnum):
     TENSORS = "tensors"
     CSV = "csv"
     JSON = "json"
+    DATASET = "dataset"
 
     # TODO: Create a FlashEnum class???
     def __hash__(self) -> int:
@@ -296,14 +298,10 @@ class DataSource(Generic[DATA_TYPE], Properties, Module):
 
             mock_dataset = typing.cast(AutoDataset, MockDataset())
             with CurrentRunningStageFuncContext(running_stage, "load_data", self):
-                load_data: Callable[[DATA_TYPE, Optional[Any]], Any] = getattr(
-                    self, DataPipeline._resolve_function_hierarchy(
-                        "load_data",
-                        self,
-                        running_stage,
-                        DataSource,
-                    )
+                resolved_func_name = DataPipeline._resolve_function_hierarchy(
+                    "load_data", self, running_stage, DataSource
                 )
+                load_data: Callable[[DATA_TYPE, Optional[Any]], Any] = getattr(self, resolved_func_name)
                 parameters = signature(load_data).parameters
                 if len(parameters) > 1 and "dataset" in parameters:  # TODO: This was DATASET_KEY before
                     data = load_data(data, mock_dataset)
@@ -319,6 +317,23 @@ class DataSource(Generic[DATA_TYPE], Properties, Module):
 
 
 SEQUENCE_DATA_TYPE = TypeVar("SEQUENCE_DATA_TYPE")
+
+
+class DatasetDataSource(DataSource):
+
+    def load_data(self, dataset: Dataset, auto_dataset: AutoDataset) -> Dataset:
+        if self.training:
+            # store a sample to infer the shape
+            parameters = signature(self.load_sample).parameters
+            if len(parameters) > 1 and AutoDataset.DATASET_KEY in parameters:
+                auto_dataset.sample = self.load_sample(dataset[0], self)
+            else:
+                auto_dataset.sample = self.load_sample(dataset[0])
+        return dataset
+
+    def load_sample(self, sample: Mapping[str, Any], dataset: Optional[Any]) -> Any:
+        # wrap everything within `.INPUT`.
+        return {DefaultDataKeys.INPUT: sample}
 
 
 class SequenceDataSource(